src/requirements.txt
flask
jinja2
flask_sqlalchemy
mariadb